ci: attempt running Kratos and Keto via GH actions services

+++ packages/keto-client-wrapper/test/keto-client-wrapper.spec.ts
@@ -41,15 +41,19 @@ describe('Keto client wrapper E2E', () => {
   };
 
   beforeAll(() => {
-    execSync(`docker-compose -f ${dockerComposeFile} up -d --wait`, {
-      stdio: 'ignore',
-    });
+    if (!process.env['CI']) {
+      execSync(`docker-compose -f ${dockerComposeFile} up -d --wait`, {
+        stdio: 'ignore',
+      });
+    }
   });
 
   afterAll(() => {
-    execSync(`docker-compose -f ${dockerComposeFile} down`, {
-      stdio: 'ignore',
-    });
+    if (!process.env['CI']) {
+      execSync(`docker-compose -f ${dockerComposeFile} down`, {
+        stdio: 'ignore',
+      });
+    }
   });
 
   beforeEach(async () => {
